refactor(LessonsTable): 拆分了组件函数，新增了函数调用
